+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Changed default log level in `mitractl` to `WARN`.
 - Stop accepting `Add(Person)` activities without `endTime` property.
 - Stop accepting proposals without `purpose` property.
+- Don't add `provider` and `receiver` properties to `Intent` objects.
 
 ### Deprecated